backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-decouple
openai
pandas
langchain 
langchain-openai 
langchain-community
pillow 
imagehash 
requests 
duckduckgo-search
google-api-python-client
